Move createExtensionDataContainer() to shared, internal location
