+++ packages/frontend-internal/src/wiring/index.ts
@@ -14,5 +14,6 @@
  * limitations under the License.
  */
 
+export { createExtensionDataContainer } from './createExtensionDataContainer';
 export { OpaqueExtensionDefinition } from './InternalExtensionDefinition';
 export { OpaqueFrontendPlugin } from './InternalFrontendPlugin';
