+++ packages/frontend-plugin-api/src/wiring/createExtensionBlueprint.test.tsx
@@ -28,8 +28,10 @@ import {
 import { createExtensionInput } from './createExtensionInput';
 import { RouteRef } from '../routing';
 import { ExtensionDefinition } from './createExtension';
-import { createExtensionDataContainer } from './createExtensionDataContainer';
-import { OpaqueExtensionDefinition } from '@internal/frontend';
+import {
+  createExtensionDataContainer,
+  OpaqueExtensionDefinition,
+} from '@internal/frontend';
 
 function unused(..._any: any[]) {}
